Use Real Data from XML Results File
The app can now read from the PACAR folder in the downloads directory as
long as the permission is granted before the app starts.
It cannot request a run-time permission through the XML Reader class
because the class is being called through the static non-activity class
DiagResultsContent, and run-time permissions require an Activity
parameter.
Also removed ITEM_MAP and replaced it completely with ITEMS because the
app kept crashing when accessing from the hashmap for some reason.